terminate Graph

+++ app/src/main/java/com/jonathan/proyectofinal/data/CognitiveExcercisesAssignment.java
@@ -10,7 +10,7 @@ public class CognitiveExcercisesAssignment {
     private String nameExcercise;
     private String descriptionExcercise;
     private int level;
-    List<Double> listScores;
+    List<Integer> listScores;
     private int bestScore;
     private String statement;
     private int rating;
@@ -18,7 +18,7 @@ public class CognitiveExcercisesAssignment {
     public CognitiveExcercisesAssignment() {
     }
 
-    public CognitiveExcercisesAssignment(String uidPatient, String uriImageExcercise, int idExcercise, String nameExcercise, String descriptionExcercise, int level, List<Double> listScores, int bestScore, String statement, int rating) {
+    public CognitiveExcercisesAssignment(String uidPatient, String uriImageExcercise, int idExcercise, String nameExcercise, String descriptionExcercise, int level, List<Integer> listScores, int bestScore, String statement, int rating) {
         this.uidPatient = uidPatient;
         this.uriImageExcercise = uriImageExcercise;
         this.idExcercise = idExcercise;
@@ -79,11 +79,11 @@ public class CognitiveExcercisesAssignment {
         this.level = level;
     }
 
-    public List<Double> getListScores() {
+    public List<Integer> getListScores() {
         return listScores;
     }
 
-    public void setListScores(List<Double> listScores) {
+    public void setListScores(List<Integer> listScores) {
         this.listScores = listScores;
     }
